feat: add YOLOv4 in image mode, the yolov4 model is trained from TAO toolkit.

+++ tis_tools/boundingbox.py
@@ -1,15 +1,16 @@
 class BoundingBox:
     def __init__(self, classID, confidence, x1, x2, y1, y2, image_width, image_height):
-        self.classID = classID
-        self.confidence = confidence
-        self.x1 = x1
-        self.x2 = x2
-        self.y1 = y1
-        self.y2 = y2
-        self.u1 = x1 / image_width
-        self.u2 = x2 / image_width
-        self.v1 = y1 / image_height
-        self.v2 = y2 / image_height
+        self.classID = int(float(classID))
+        self.confidence = float(confidence)
+        self.x1 = int(float(x1))
+        self.x2 = int(float(x2))
+        self.y1 = int(float(y1))
+        self.y2 = int(float(y2))
+        
+        self.u1 = float(x1 / image_width)
+        self.u2 = float(x2 / image_width)
+        self.v1 = float(y1 / image_height)
+        self.v2 = float(y2 / image_height)
     
     def box(self):
         return (self.x1, self.y1, self.x2, self.y2)
